+++ amd/src/calendar_actions.js
@@ -172,8 +172,11 @@ define([
                         CalendarViewManager.stopLoading(root);
                     } else {
                         clean();
-                        // Redirect to bookings view
-                        location.href = M.cfg.wwwroot + '/local/booking/view.php?courseid=' + course;
+                        // Redirect to the report page instead of view.php
+                        window.location.href = M.cfg.wwwroot +
+                            // '/local/booking/booking_report.php?bookingid=' + response.bookingid +
+                            // '&courseid=' + course + '&studentid=' + studentid;
+                            '/local/booking/view.php?courseid=' + course;
                     }
                     return;
                 })